src/requirements.txt
# ─── 웹 스크래핑 ───────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ─── 언어 감지 ─────────────────────────────────
langdetect>=1.0.9

# ─── AI (시나리오/이미지/TTS) ──────────────────
openai>=1.30.0

# ─── Edge TTS (무료 로컬 TTS) ─────────────────
edge-tts>=6.1.9

# ─── 영상 합성 ─────────────────────────────────
moviepy>=1.0.3
ffmpeg-python>=0.2.0
Pillow>=10.0.0

# ─── 오디오 처리 ───────────────────────────────
pydub>=0.25.1

# ─── 유튜브 업로드 ─────────────────────────────
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ─── 이미지 유사도 (캐시 매처) ─────────────────
scikit-image>=0.22.0
numpy>=1.26.0

# ─── 환경변수 ──────────────────────────────────
python-dotenv>=1.0.0

# ─── GUI ───────────────────────────────────────
# PyQt5>=5.15.0   # 유료 라이선스 필요 시 PyQt6 대체 가능
PyQt6>=6.6.0

# ─── 스케줄러 ──────────────────────────────────
schedule>=1.2.0

# ─── 빌드 ──────────────────────────────────────
pyinstaller>=6.0.0
